Consolidate types
