+++ packages/react-day-picker/src/components/Caption/Caption.tsx
@@ -4,10 +4,10 @@ import { isSameMonth } from 'date-fns';
 
 import { useDayPicker } from '../../contexts/DayPicker';
 import { useNavigation } from '../../contexts/Navigation';
-import { MonthChangeEventHandler } from '../../types';
 import { MonthsDropdown } from '../MonthsDropdown';
 import { Navigation } from '../Navigation';
 import { YearsDropdown } from '../YearsDropdown';
+import { MonthChangeEventHandler } from '../../types/EventHandlers';
 
 /** Represent the props of the [[Caption]] component. */
 export interface CaptionProps {
