+++ packages/react-day-picker/src/components/MonthsDropdown/MonthsDropdown.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { isSameYear, setMonth as setDateMonth, startOfMonth } from 'date-fns';
 
 import { useDayPicker } from '../../contexts/DayPicker';
-import { MonthChangeEventHandler } from '../../types';
+import { MonthChangeEventHandler } from '../../types/EventHandlers';
 
 /** The props for the [[MonthsDropdown]] component. */
 export interface MonthsDropdownProps {
