+++ packages/react-day-picker/src/components/YearsDropdown/YearsDropdown.tsx
@@ -2,9 +2,8 @@ import React from 'react';
 
 import { setYear, startOfYear } from 'date-fns';
 
-import { MonthChangeEventHandler } from '../../types';
-
 import { useDayPicker } from '../../contexts/DayPicker';
+import { MonthChangeEventHandler } from '../../types/EventHandlers';
 
 /**
  * The props for the [[YearsDropdown]] component.
