+++ packages/react-day-picker/src/contexts/DayPicker/labels/labelDay.ts
@@ -1,10 +1,10 @@
 import { format } from 'date-fns';
 
-import { DayLabelFormatter } from '../../../types';
+import { DayLabel } from '../../../types';
 
 /**
  * The default ARIA label for the day button.
  */
-export const labelDay: DayLabelFormatter = (day, options): string => {
+export const labelDay: DayLabel = (day, options): string => {
   return format(day, 'do MMMM (EEEE)', options);
 };
