+++ packages/react-day-picker/src/contexts/DayPicker/labels/labelNext.ts
@@ -1,8 +1,8 @@
-import { NavButtonLabelFormatter } from '../../../types';
+import { NavButtonLabel } from '../../../types';
 
 /**
  * The default ARIA label for next month button in navigation
  */
-export const labelNext: NavButtonLabelFormatter = (): string => {
+export const labelNext: NavButtonLabel = (): string => {
   return 'Go to next month';
 };
