+++ packages/react-day-picker/src/contexts/DayPicker/labels/labelPrevious.ts
@@ -1,8 +1,8 @@
-import { NavButtonLabelFormatter } from '../../../types';
+import { NavButtonLabel } from '../../../types';
 
 /**
  * The default ARIA label for previous month button in navigation
  */
-export const labelPrevious: NavButtonLabelFormatter = (): string => {
+export const labelPrevious: NavButtonLabel = (): string => {
   return 'Go to previous month';
 };
