+++ packages/react-day-picker/src/hooks/useInput/useInput.ts
@@ -4,12 +4,12 @@ import { differenceInCalendarDays, format as _format, parse } from 'date-fns';
 import enUS from 'date-fns/locale/en-US';
 
 import { parseFromToProps } from '../../contexts/DayPicker/utils';
+import { isValidDate } from './utils/isValidDate';
 import {
   DayClickEventHandler,
-  DayPickerProps,
   MonthChangeEventHandler
-} from '../../types';
-import { isValidDate } from './utils/isValidDate';
+} from '../../types/EventHandlers';
+import { DayPickerProps } from '../../types/DayPicker';
 
 /** Represent the props to attach to the input field. */
 export interface UseInputFieldProps {
