--- packages/react-day-picker/src/types/ClassNames.ts
@@ -1,4 +1,0 @@
-import { StyledElement } from './StyledElement';
-
-/** The class names of each element. */
-export type ClassNames = Partial<StyledElement<string>>;
