--- packages/react-day-picker/src/types/DateFormatter.ts
@@ -1,7 +1,0 @@
-import { Locale } from 'date-fns';
-
-/** Represents a function to format a date. */
-export type DateFormatter = (
-  date: Date,
-  options?: { locale?: Locale }
-) => React.ReactNode;
