--- packages/react-day-picker/src/types/DayClickEventHandler.ts
@@ -1,10 +1,0 @@
-import { ModifierStatus } from './Modifiers';
-
-/**
- * Represent the event handler when a day is clicked.
- */
-export type DayClickEventHandler = (
-  day: Date,
-  modifiers: ModifierStatus,
-  e: React.MouseEvent
-) => void;
