--- packages/react-day-picker/src/types/DayFocusEventHandler.ts
@@ -1,10 +1,0 @@
-import { ModifierStatus } from './Modifiers';
-
-/**
- * Represent the event handler when a day is focused.
- */
-export type DayFocusEventHandler = (
-  day: Date,
-  modifiers: ModifierStatus,
-  e: React.FocusEvent | React.KeyboardEvent
-) => void;
