--- packages/react-day-picker/src/types/DayKeyboardEventHandler.ts
@@ -1,10 +1,0 @@
-import { ModifierStatus } from './Modifiers';
-
-/**
- * Represent the event handler when a day gets a keyboard event.
- */
-export type DayKeyboardEventHandler = (
-  day: Date,
-  modifiers: ModifierStatus,
-  e: React.KeyboardEvent
-) => void;
