+++ packages/react-day-picker/src/types/DayPickerSingle.ts
@@ -1,9 +1,7 @@
-import { DayPickerProps } from './DayPickerProps';
-import { SelectSingleEventHandler } from './SelectSingleEventHandler';
+import { DayPickerProps } from './DayPicker';
+import { SelectSingleEventHandler } from './EventHandlers';
 
-/**
- * The props for the [[DayPicker]] component when using `mode="single"`.
- */
+/** The props for the [[DayPicker]] component when using `mode="single"`. */
 export interface DayPickerSingleProps extends DayPickerProps {
   mode: 'single';
   /** The selected day. */
@@ -13,6 +11,7 @@ export interface DayPickerSingleProps extends DayPickerProps {
   /** Make the selection required. */
   required?: boolean;
 }
+
 /** Returns true when the props are of type [[DayPickerSingle]]. */
 export function isDayPickerSingle(
   props: DayPickerProps
