--- packages/react-day-picker/src/types/DayTouchEventHandler.ts
@@ -1,10 +1,0 @@
-import { ModifierStatus } from './Modifiers';
-
-/**
- * Represent the event handler when a day gets a touch event.
- */
-export type DayTouchEventHandler = (
-  day: Date,
-  modifiers: ModifierStatus,
-  e: React.TouchEvent
-) => void;
