+++ packages/react-day-picker/src/types/Labels.ts
@@ -1,46 +1,35 @@
-/**
- * Map of functions to translate ARIA labels for the relative elements.
- */
+import { Locale } from 'date-fns';
+import { ModifierStatus } from '../types';
+
+/** Map of functions to translate ARIA labels for the relative elements. */
 export type Labels = {
   labelMonthDropdown: () => string;
   labelYearDropdown: () => string;
-  labelNext: NavButtonLabelFormatter;
-  labelPrevious: NavButtonLabelFormatter;
-  labelDay: DayLabelFormatter;
-  labelWeekday: WeekdayLabelFormatter;
-  labelWeekNumber: WeekNumberLabelFormatter;
+  labelNext: NavButtonLabel;
+  labelPrevious: NavButtonLabel;
+  labelDay: DayLabel;
+  labelWeekday: WeekdayLabel;
+  labelWeekNumber: WeekNumberLabel;
 };
 
-import { Locale } from 'date-fns';
-
-import { ModifierStatus } from '../types';
-
-/**
- * Represent a function to format the ARIA label for the [[Day]] component.
- */
-export type DayLabelFormatter = (
+/** Return the ARIA label for the [[Day]] component. */
+export type DayLabel = (
   day: Date,
   modifiers: ModifierStatus,
   options?: { locale?: Locale }
 ) => string;
 
-/**
- * Represent a function to format the ARIA label for the "next month" / "prev
- * month" buttons in the navigation.
- */
-export type NavButtonLabelFormatter = (
+/** Return the ARIA label for the "next month" / "prev month" buttons in the navigation.*/
+export type NavButtonLabel = (
   month?: Date,
   options?: { locale?: Locale }
 ) => string;
 
-/** Represent a function to format the ARIA label for the Head component.*/
-export type WeekdayLabelFormatter = (
-  day: Date,
-  options?: { locale?: Locale }
-) => string;
+/** Return the ARIA label for the Head component.*/
+export type WeekdayLabel = (day: Date, options?: { locale?: Locale }) => string;
 
-/** Represent a function to format the ARIA label of the week number.*/
-export type WeekNumberLabelFormatter = (
+/** Return the ARIA label of the week number.*/
+export type WeekNumberLabel = (
   n: number,
   options?: { locale?: Locale }
 ) => string;
