--- packages/react-day-picker/src/types/MonthChangeEventHandler.ts
@@ -1,4 +1,0 @@
-/**
- * Represent the event handler when a month is changed in the calendar.
- */
-export type MonthChangeEventHandler = (month: Date) => void;
