+++ packages/react-day-picker/src/types/index.ts
@@ -1,29 +1,11 @@
-export * from './ClassNames';
-export * from './Components';
-export * from './DateFormatter';
-export * from './DayClickEventHandler';
-export * from './DayFocusEventHandler';
-export * from './DayKeyboardEventHandler';
-export * from './DayMouseEventHandler';
-export * from './DayPickerProps';
-export * from './DayPickerCustomProps';
-export * from './DayPickerMultipleProps';
-export * from './DayPickerProps';
-export * from './DayPickerRangeProps';
-export * from './DayPickerSingleProps';
-export * from './DayTouchEventHandler';
+export * from './DayPicker';
+export * from './DayPickerCustom';
+export * from './DayPickerMultiple';
+export * from './DayPicker';
+export * from './DayPickerRange';
+export * from './DayPickerSingle';
 export * from './Formatters';
-export * from './isArrayOfDates';
-export * from './isDateType';
 export * from './Labels';
 export * from './Matchers';
 export * from './Modifiers';
-export * from './MonthChangeEventHandler';
-export * from './SelectMultipleEventHandler';
-export * from './SelectRangeEventHandler';
-export * from './SelectSingleEventHandler';
-export * from './StyledComponentProps';
-export * from './StyledElement';
 export * from './Styles';
-export * from './WeekNumberClickEventHandler';
-export * from './WeekNumberFormatter';
