--- packages/react-day-picker/src/types/isDateType.ts
@@ -1,6 +1,0 @@
-import { isDate } from 'date-fns';
-
-/** Returns true if `value` is a Date type. */
-export function isDateType(value: unknown): value is Date {
-  return isDate(value);
-}
